requirements.txt
numpy==1.24.4
torch==2.1.0
tensorboard==2.14.0
tqdm==4.66.1
transformers==4.37.0
datasets==2.15.0
openai==1.6.1
sentencepiece==0.1.99
nltk==3.8.1
spacy==3.7.2
gensim==4.3.2
scikit-learn==1.3.0
sentence-transformers==2.2.2

stefutils==0.37.4  # general utilities
sacremoses==0.1.1  # for detokenization, i.e. joining tokens back into text
tiktoken==0.4.0  # for counting openai completion tokens
# for computing string pair edit distance,
#   used in comparing LLM-generated sentences w/ sentences in prompt to log edge cases
Levenshtein==0.23.0
seqeval==1.2.2  # for NER sequence/chunk-wise eval metrics
nervaluate==0.1.8  # like `seqeval`, but supports partial match
